// web-view/src/main/todo/src/app/angular-client/model/user.ts
/**
 * ToDo-Web-API
 * ToDo list sample app web frontend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';
import { SuperUpdateIdPageable } from './superUpdateIdPageable';


export interface User { 
    /**
     * User name
     */
    name: string;
    /**
     * User last name
     */
    lastName: string;
    /**
     * User login
     */
    login: string;
    /**
     * User password
     */
    password?: string;
    /**
     * Last modification date of the task
     */
    updateDate?: string;
    /**
     * User roles
     */
    roles?: Array<Role>;
    id?: number;
    version: number;
    /**
     * Total number of pages
     */
    totalPages?: number;
    /**
     * Current page number
     */
    currentPage?: number;
}